Separate trait from concrete implementation
