+++ client/src/main/scala/com.gu.contentapi.client/ContentApiClient.scala
@@ -1,15 +1,11 @@
 package com.gu.contentapi.client
 
-import java.io.IOException
-import java.util.concurrent.TimeUnit
 import com.gu.contentapi.client.model._
 import com.gu.contentapi.client.model.v1._
-import com.gu.contentapi.client.utils.QueryStringParams
-import com.gu.contentapi.buildinfo.CapiBuildInfo
-import okhttp3._
-import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.util.Try
 import com.gu.contentapi.client.thrift.ThriftDeserializer
+import com.gu.contentapi.client.utils.QueryStringParams
+import scala.concurrent.{ExecutionContext, Future}
+import scala.util.Try
 
 case class GuardianContentApiError(httpStatus: Int, httpMessage: String, errorResponse: Option[ErrorResponse] = None) extends Exception(httpMessage)
 
@@ -137,43 +133,3 @@ trait ContentApiClient {
     }
 
 }
-
-class GuardianContentClient(val apiKey: String) extends ContentApiClient {
-
-  val userAgent = "content-api-scala-client/"+CapiBuildInfo.version
-  val targetUrl = "https://content.guardianapis.com"
-
-  private val http = new OkHttpClient.Builder()
-    .connectTimeout(1, TimeUnit.SECONDS)
-    .readTimeout(2, TimeUnit.SECONDS)
-    .followRedirects(true)
-    .connectionPool(new ConnectionPool(10, 60, TimeUnit.SECONDS))
-    .build()
-
-  def get(url: String, headers: Map[String, String])(implicit context: ExecutionContext): Future[HttpResponse] = {
-
-    val reqBuilder = new Request.Builder().url(url)
-    val req = headers.foldLeft(reqBuilder) {
-      case (r, (name, value)) => r.header(name, value)
-    }
-
-    val promise = Promise[HttpResponse]()
-
-    http.newCall(req.build()).enqueue(new Callback() {
-      override def onFailure(call: Call, e: IOException): Unit = promise.failure(e)
-      override def onResponse(call: Call, response: Response): Unit = {
-        promise.success(HttpResponse(response.body().bytes, response.code(), response.message()))
-      }
-    })
-
-    promise.future
-  }
-
-  /** Shutdown the client and clean up all associated resources.
-    *
-    * Note: behaviour is undefined if you try to use the client after calling this method.
-    */
-  def shutdown(): Unit = http.dispatcher().executorService().shutdown()
-
-}
-
